+++ packages/react-tinacms-github/package.json
@@ -18,6 +18,7 @@
     "final-form-submit-listener": "^1.0.0",
     "js-cookie": "^2.2.1",
     "styled-components": "^5.0.1",
-    "tinacms": "^0.18.2"
+    "tinacms": "^0.18.2",
+    "@tinacms/react-forms": "^0.1.0"
   }
 }
