+++ packages/react-tinacms-github/src/open-authoring-ui/components/AsyncButton.tsx
@@ -1,0 +1,31 @@
+import { useCallback, useState } from 'react'
+import { LoadingDots } from '@tinacms/react-forms'
+import { Button } from '@tinacms/styles'
+
+interface ButtonProps {
+  name: string
+  action(): Promise<void>
+  primary: boolean
+}
+
+export const AsyncButton = ({ name, primary, action }: ButtonProps) => {
+  const [submitting, setSubmitting] = useState(false)
+
+  const onClick = useCallback(async () => {
+    setSubmitting(true)
+    await action()
+    setSubmitting(false)
+  }, [action, setSubmitting])
+
+  return (
+    <Button
+      primary={primary}
+      onClick={onClick}
+      busy={submitting}
+      disabled={submitting}
+    >
+      {submitting && <LoadingDots />}
+      {!submitting && name}
+    </Button>
+  )
+}
